[Vulkan] Added feature hook generation for commands - Backend library drives the hook definitions, i.e. the abstraction, the backend project (Vulkan, DX12) drive how it's translated to that (See FeatureProxies.h) - Added simple Delegate type, fast single indirection function pointers - Added some work-in-progress IL, such as instructions, basic blocks and emitters   (super early)
Debating if it's easier to just translate the entire program to the generalized AST, than to specialize it per backend. I am not talking bout translating all instructions, that's a waste of time, but just enough of them. Then based on the modified basic blocks, you could apply the changes. Many things to try!