Added Libraries.Message, a fast message allocation and storage library. Supports inline arrays and indirections, and three schemas: - Static   The given message stream only contains a single type, no header allocation. - Dynamic   The given message only contains a single type of variable size, small header allocation. - Ordered   The given message stream supports any given type with variable sizes, moderate header allocation.
The majority of all streams, from a given feature, will be static or dynamic. This combined with the batched nature, should result in rapid allocation and visitation times.

All messages are inline and flat in memory, any data referenced by a message will be in the byte boundary of the message. Due to this flat nature, serialization of messages across network or language boundaries is not required, as long as the memory layout is consistent.

Looked at various libraries such as FlatBuffers, Protobuf, etc... however, were not seen as a good fit.