[Vulkan] Added simple SPIRV parsing for program abstraction - Added SPIRV-Headers - Fixed deep copy generator memcpy usage
I considered using SPIRV-Tools as the proof-of-concept uses, however, from my experience, I find it simpler to just write a custom parser. The format itself is incredibly simple, and easy to manipulate and reason about.

If need be, it would be easy to use SPIRV-Tools in the future, as everything will go through this thin abstraction IL.