[Vulkan, --] Functional shader exporting via the messaging API, uniform and universal message writing between host and device, with no translation requirement. Added unit test which demonstrates the usage.
To create a shader compatible message export, simply define a schema such as (full spec found in documentation):
	<schema>
		<shader-export name="MessageName">
			<field name="fieldName" type="uint32" bits="16"/>
		</shader-export>
	</schema>

To export the message from a shader, implement the IShaderFeature interface and register a new export with IShaderExportHost.
With that, during IL writing, export as following:
	MessageName::ShaderExport msg;
	msg.sguid = emitter.UInt(32, 0);
	msg.fieldName = emitter.UInt(32, kValue);
	emitter.Export(exportID, msg);

Packing, exporting, streaming and message passing is handled by the backend in a performant manner.

[Vulkan.Tests]
- Added validation test for negative UAV writes, writes a series of validation messages that are then validated CPU side.
- Added Khronos validation layer support

[Vulkan]
- Added shader export descriptor allocator, responsible for descriptor mapping and lifetimes. During recording
of command buffer we do not know where it is yet to be submitted. So the descriptors are updated in a latent fashion.
- Added shader export allocators, all instrumentation is done in "segments", which are typically a chunk of submitted command buffers
on a given queue. A shared buffer is used to host the atomically incremented counters, whereas the individual export streams
are hosted separately. The separation of streams greatly reduces the size of each exported message, given that it frees up some of the bits
for small messages. Currently the streams are hosted in mappable memory, however this may change. Counters are mirrored allocations, given
that they are atomically incremented (would be far too slow otherwise).
- First implementation of the export streamer, handles command buffer injection, descriptor compatability, and segment processing.
- Instrumentation now operates on "ShaderModuleInstrumentationKey", pipeline layouts may change between
pipelines with the same shaders bound. During shader compilation, we do not know this, and simply using the last set
may result in incompatabilities. Instead, this key is composed of the enabled feature bit set (as it is today),
and the pipeline layout bind count. This ensures we maintain compatibility with all pipeline layouts, however may
result in additional compilations. Alternatively, we may inject a set at the beginning, and then offset all remaining sets. However,
this is greatly complicated when the max set limit is reached, and may yield excessive ping ponging between
layouts with the extra record injected, and those without. Additionally, it goes against the design philosphy of excessively
changing the base behaviour, as it would require shader injection even when the layer is not "active". The initial
paper that proposed GPU based validation, suggested to bind this set at a sensible offset, however to pad all missing
sets with dummy sets. This was seen as too excessive of an overhead, and does not solve the initial issue of pipeline compatability.
- Added support for the new compilation model to the instrumentation controller
- Moved fences to reference counted objects, for fence sharing. Added commit ids to the fence objects to ensure that the right value
is being waited for, in case the user resets the fence.
- Added hooks for fence creation, destruction, querying and resetting.
- Moved pipeline layouts to reference counted objects, injects shader export record if possible (hasn't reached device limit).
- Added hooks for pipeline layout creation and descriptor binding
- Added queue states, hosts shader export states and provides temporary command buffers for immediate work.
- Added hooks for queue waiting, acts as sync points for shader export processing.
- SpvModule now writes the shader export descriptor records
- Added support for array and image spirv types
- Added support for spriv capabilities, and parsing of the physical section layouts
- Added support for the IL export instruction, writes to the bound export record

- Added lifetime bound objects to command buffer objects
- Added device allocation helpers, and mirror allocations when device residency is required
- Command buffer dispatch generation exports proc fetches for whitelisted commands
- Changed the directory structure slightly
- Added Remove helper for tracked objects
- Added internal dispatch table patching for validation layers
- Shader compiler was not respecting the enabled feature bit set
- Device was not advancing the link info, fixed
- Enable AfterBind during device creation (needs a proper system)
- Fixed VMA initialization, route proc addresses to VMA

[Backend]
- Added type info and enumeration to the shader export host
- Added export counter type

[Bridge]
- Fixed unordered message stream commits

[Common]
- Added indirect and trivial object pools for object recycling
- Moved ReferenceObject to Common
- Fixed (embarrasing) ALLOCA_ARRAY bug with macro encapsulation, had this on an interview test once

[Build]
- Fixed MSVC toolchain errors
- Added offline working for third party pulling (Offline mode "INSTALL_THIRD_PARTY 0" improves compilation speeds significantly)

[Git]
- Extended .gitingore to most common IDE's