[Bridge, --] Second iteration on network architecture, added support for multiple injected applications, attachable (network wise) from any endpoint.
The previous iteration simply created a server per application, which of course does not scale beyond a single application as the discovery was tied to a single port which the succeeding applications could not allocate.

To avoid this issue, a unique instance of the so call "host resolver" server runs on demand, which is responsible for managing all the injected applications. This is both automatic discovery of said applications, and managing the remote client that wishes to connect to one of them.

The remote client then first performs discovery through the resolver, and then wakes a server on an application of interest. Once the server is ready, the details are sent back to the remote client. From there, the same architecture as last is used. An alternative was to simply start a server per application from the get go, however this is a lot of network pollution.

Functionality wise this is pretty solid (lots of bugs, particularly regarding lost connections), however, there are obvious security implications to this, which has yet to be addressed.

[Bridge]
- Added HostServerBridge / RemoteClientBridge, replaces NetworkBridge with further specializations.
- Added host resolve schema for asio to message translation
- Added AsioHostResolverServer, responsible for managing all local servers and proxying with remote clients
- Added AsioHostServer, on-demand server for remote client requests. Handled through the resolver server.
- Added AsioRemoteClient, interfaces with the host server through the resolve server.
- Added guid to socket handlers, allows for safe referencing.
- Moved socket buffer to heap
- Removed generic endpoint interface
- AsioClient must stop io service during shutdown
- Added AsioAsyncRunner, simple async worker thread

[HostResolver]
- Added host resolving service, manages handshakes and discovery between remote clients and local servers.

[Backend]
- Networked environment initializes host resolver if not present
- Changed to HostServerBridge instead of NetworkedBridge
- Added application name to environment info

[Vulkan]
- Added logging to instrumentation controller
- Instance passes application name to environment info

[Message]
- Fixed stomp caused by incorrect offsets for sub streams
- Added missing ALIGN_PACK to messages
- Added raw Set to MessageString
- Fixed IsEmpty for sub streams

[Common]
- Added GlobalUID class, full Win32 implementation, partial Linux implementation.
- Added EventHandler, simple collection of invocables
- Added CurrentExecutableName
- Do not add exception handler if debugger is present