[Build] Added workaround for CI compilation bug
The issue is that cmake --build operates on the BUILD_ALL project instead of the solution, where the solution may define extra dependencies that are not inside the individual project files. This a cmake bug.

See issue in:
https://gitlab.kitware.com/cmake/cmake/-/issues/18026

The workaround is instead to invoke msbuild. This changelist is a hypothetical fix to the problem.