[Vulkan, DX12, UIX] Added versioning system, initially used for PUID to resource lookups
Simply doing a 1:1 PUID to resource lookup is innacurate, by the time a PUID has arrived the given resource underneath may be invalidated or replaced. To combat this a lightweight versioning system was introduced, which contains these lookups in branches.

Backends are not responsible for managing said branches, instead they just update the current "status". Branches are created during segmentation points, f.x. submission, and all new resources land in the next branch. This also means that message streams have the concept of a version now, which is 1:1 to the current head index of the versioning system.

The listener (UIX) keeps track of the current branches, and performs the appropriate mapping to ensure the fetched resource is exactly what was invoked during the command list segment. Additionally, to not accumulate infinite space, the branches are collapsed when the submissions are known as done.

[UIX]
- Added versioning service, handles branch tracking, collection, population and collapsing. Additionally, provides lookup.
- Added Resource states, contains basic information about resources
- Added reosurce token wrapper

[Backend]
- Added ResourceVersion, GetVersionSummarization, VersionSummarization, VersionBranch, VersionCollapse messages.

[Message]
- Added version ids to message streams

[Bridge]
- Added support for streaming stream versions
- Rewrote memory bridge commits, must be in order. This will be revisited in the future.

[ResourceBounds]
- Compose detail resource from versioning

[DX12, Vulkan]
- Added versioning controller. Handles version commits, branching, summarization, and collapses.
- Added image debug naming
- Added buffer debug naming

[DX12]
- Added resource state tracking
- Added resource debug naming