[DX12] Bootstrapper jump block trap guards
This was an interesting issue to track down, what appears to happen is that the (detouring) jump block gets overwritten on each hooked call, caused by GameManager32.exe. I'm not entirely sure why it's doing that or what the purpose is, maybe to restore some internal hooking state, either way, no way.

DetourJmpSectionTrapGuard checks if the jmp instruction has changed, if so, write it back again.

- Fixed a race condition on deferred initialization